requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-rest-knox==4.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
Markdown==3.4.1
mysqlclient==2.1.1
Pillow==9.4.0
pycparser==2.21
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
